[ORGANIZE] Move wallet related things to src/wallet/
Move wallet related files to src/wallet/
