+++ src/qt/addresstablemodel.cpp
@@ -8,7 +8,7 @@
 #include "walletmodel.h"
 
 #include "base58.h"
-#include "wallet.h"
+#include "wallet/wallet.h"
 
 #include <QFont>
 #include <QDebug>
