+++ src/qt/paymentserver.cpp
@@ -12,7 +12,7 @@
 #include "chainparams.h"
 #include "ui_interface.h"
 #include "util.h"
-#include "wallet.h"
+#include "wallet/wallet.h"
 
 #include <cstdlib>
