+++ src/qt/signverifymessagedialog.cpp
@@ -12,7 +12,7 @@
 
 #include "base58.h"
 #include "init.h"
-#include "wallet.h"
+#include "wallet/wallet.h"
 
 #include <string>
 #include <vector>
