+++ src/qt/walletmodeltransaction.cpp
@@ -4,7 +4,7 @@
 
 #include "walletmodeltransaction.h"
 
-#include "wallet.h"
+#include "wallet/wallet.h"
 
 WalletModelTransaction::WalletModelTransaction(const QList<SendCoinsRecipient> &recipients) :
     recipients(recipients),
