+++ src/rpcrawtransaction.cpp
@@ -16,7 +16,7 @@
 #include "script/standard.h"
 #include "uint256.h"
 #ifdef ENABLE_WALLET
-#include "wallet.h"
+#include "wallet/wallet.h"
 #endif
 
 #include <stdint.h>
